Do not follow links when validating staged files
When updating 3rd-party bundles, swupd validates the files to be updated
to make sure there were no changes in file permissions that could be
risky for the system, it uses stat to get the file permissions. The
problem is that stat follows links, and while the files are in staging,
the links are really pointing to invalid locations so stat will always
fail.

This commit fixes the problem by not following links when validating
file permissions.

Signed-off-by: Castulo Martinez <castulo.martinez@intel.com>